fix: lists do not update after site generation
Due to the site being generated in Nuxt full-static mode, the fetch hook
only runs during the generation step and the result stays cached.

I wanted to be able to at least refresh the lists or to fetch it again
when navigating.

To refresh the lists, the list headings now come with a button for their
own list. This also meant splitting the list fetching into their own
methods.

To refresh the lists when navigating, I used the `fetchKey` property and
return a key based on the time to make sure it always changes on load.